Adjust a covariance matrix for Common Method Variance (CMV)

sqp_cmv_cov accepts a covariance matrix, a covariance data frame or a covariance tibble from sqp_covariance and adjusts the coefficients of the variables specified in ... with the reliability and validity coefficients given by get_estimates. All variables specified in ... must be present in both x and sqp_data. Optionally, you can supply the cmv coefficient in the argument cmv.

sqp_cmv_cov(x, sqp_data, ..., original_data, cmv = NULL)

Arguments

x
a covariance matrix, a covariance data frame or a covariance tibble given by sqp_covariance.

sqp_data
a data frame or tibble of class sqp given by get_estimates.

...
two or more variables present in both x and sqp_data. Can be both in bare unquoted names or as character strings.

original_data
A data frame containing the original data for variables specified in .... These are the original columns of the variables from the covariance matrix supplied in x.

cmv
an optional numeric vector of length 1 which contains the CMV coefficient of the variables specified in .... This argument is left available if the user has reasons to input their own CMV. By default, it is set to NULL and it is calculated internally.

Value

the same matrix supplied in x but as a tibble with the covariance coefficients of the variables supplied in ... adjusted for their common method variance.

If x had row names, they're moved as a column called 'rowname'. If it doesnt have row names or a column named 'rowname' a new column is created with the column names to mimic a covariance matrix.

See also

Examples

set.seed(2131)
library(tibble)

original_df <- as.data.frame(matrix(rnorm(100, sd = 50), nrow = 20))
cov_tibble <- sqp_covariance(original_df)

# Toy dataset
sqp_df <-
 tibble(question = paste0("V", 1:5),
 quality = c(0.2, 0.3, 0.5, 0.6, 0.9),
 reliability = c(NA, 0.4, 0.5, 0.5, 0.7),
 validity = c(NA, NA, 0.6, 0.7, 0.8))

sqp_df <- structure(sqp_df, class = c(class(sqp_df), "sqp"))

# Show that when y is not from sqp class, there's an error

# Original covariance matrix with new diagonal
cov_tibble

#> # A tibble: 5 x 6
#>   rowname    V1    V2     V3     V4    V5
#>   <chr>   <dbl> <dbl>  <dbl>  <dbl> <dbl>
#> 1 V1      2411.  129.   608.   445.  229.
#> 2 V2       129. 1653.   784.  -282.  618.
#> 3 V3       608.  784.  2446. -1146. -574.
#> 4 V4       445. -282. -1146.  2772. -181.
#> 5 V5       229.  618.  -574.  -181. 3531.

# Coefficient of covariance changes
# when adjusting for common method variance
sqp_cmv_cov(cov_tibble, sqp_df, V4, V5, original_data = original_df)

#> # A tibble: 5 x 6
#>   rowname    V1    V2     V3     V4    V5
#>   <chr>   <dbl> <dbl>  <dbl>  <dbl> <dbl>
#> 1 V1      2411.  129.   608.   445.  229.
#> 2 V2       129. 1653.   784.  -282.  618.
#> 3 V3       608.  784.  2446. -1146. -574.
#> 4 V4       445. -282. -1146.  2772. -634.
#> 5 V5       229.  618.  -574.  -634. 3531.

# The V5*V4 from both the upper/lower triangles
# covariance matrix changed from -181 to -634